// backend/src/helpers/interfaces.ts
export interface IDb {
	Patient: any;
	ChiefComplaint: any;
	PastMedicalHistory: any;
	PastDentalHistory: any;
	DrugAllergy: any;
	GeneralExamination: any;
	LocalExamination: any;
	ClinicalDiagnosis: any;
	Investigation: any;
	Diagnosis: any;
	TreatmentPlan: any;
	Medicine: any;
	FollowUp: any;
	Favourite: any;
}